refactor(ecs): move Entity its own file
