+++ core/include/cubos/core/ecs/entity/manager.hpp
@@ -1,52 +1,18 @@
 /// @file
-/// @brief Class @ref cubos::core::ecs::EntityManager and related types.
+/// @brief Class @ref cubos::core::ecs::EntityManager.
 /// @ingroup core-ecs
 
 #pragma once
 
-#include <bitset>
-#include <cstdint>
 #include <queue>
 #include <set>
 #include <unordered_map>
 #include <vector>
 
+#include <cubos/core/ecs/entity/entity.hpp>
+
 namespace cubos::core::ecs
 {
-    /// @brief Identifies an entity.
-    ///
-    /// When serializing/deserializing, if there's a
-    /// data::old::SerializationMap<Entity, std::string> in the context, it will be used to
-    /// (de)serialize strings representing the entities. Otherwise, the identifiers will be
-    /// (de)serialized as objects with two fields: their index and their generation.
-    ///
-    /// @ingroup core-ecs
-    class Entity
-    {
-    public:
-        /// @brief Type used to store which components an entity has.
-        using Mask = std::bitset<CUBOS_CORE_ECS_MAX_COMPONENTS + 1>;
-
-        Entity();
-
-        /// @brief Constructs an entity from an index and a generation.
-        /// @param index Entity index.
-        /// @param generation Entity generation.
-        Entity(uint32_t index, uint32_t generation);
-
-        Entity(const Entity&) = default;
-        Entity& operator=(const Entity&) = default;
-        bool operator==(const Entity& /*entity*/) const;
-        bool operator!=(const Entity& /*entity*/) const;
-
-        /// @brief Checks if the entity is null, a special value returned on errors.
-        /// @return Whether the entity is null.
-        bool isNull() const;
-
-        uint32_t index;      ///< Index in storages.
-        uint32_t generation; ///< Allows us to detect if the entity has been removed.
-    };
-
     /// @brief Holds and manages entities and their component masks.
     ///
     /// Used internally by @ref World.
@@ -144,17 +110,3 @@ namespace cubos::core::ecs
         std::unordered_map<Entity::Mask, std::set<uint32_t>> mArchetypes; ///< Cache archetype entity indices.
     };
 } // namespace cubos::core::ecs
-
-namespace std
-{
-    // Add hash function for Entity, so that it can be used as a key in an unordered_map.
-
-    template <>
-    struct hash<cubos::core::ecs::Entity>
-    {
-        inline std::size_t operator()(const cubos::core::ecs::Entity& k) const
-        {
-            return hash<uint32_t>()(k.index);
-        }
-    };
-} // namespace std
