+++ core/src/cubos/core/ecs/entity/manager.cpp
@@ -1,94 +1,6 @@
-#include <cubos/core/data/old/deserializer.hpp>
-#include <cubos/core/data/old/serialization_map.hpp>
-#include <cubos/core/data/old/serializer.hpp>
 #include <cubos/core/ecs/entity/manager.hpp>
 
 using namespace cubos::core::ecs;
-using namespace cubos::core::data;
-
-template <>
-void cubos::core::data::old::serialize<Entity>(Serializer& ser, const Entity& obj, const char* name)
-{
-    if (ser.context().has<SerializationMap<Entity, std::string>>())
-    {
-        if (obj.isNull())
-        {
-            ser.write("null", name);
-            return;
-        }
-
-        auto& map = ser.context().get<SerializationMap<Entity, std::string>>();
-        ser.write(map.getId(obj), name);
-    }
-    else
-    {
-        ser.beginObject(name);
-        ser.write(obj.index, "index");
-        ser.write(obj.generation, "generation");
-        ser.endObject();
-    }
-}
-
-template <>
-void cubos::core::data::old::deserialize<Entity>(Deserializer& des, Entity& obj)
-{
-    if (des.context().has<SerializationMap<Entity, std::string>>())
-    {
-        std::string name;
-        des.read(name);
-        if (name == "null")
-        {
-            obj = Entity();
-            return;
-        }
-
-        auto& map = des.context().get<SerializationMap<Entity, std::string>>();
-        if (map.hasId(name))
-        {
-            obj = map.getRef(name);
-        }
-        else
-        {
-            CUBOS_WARN("No such entity '{}'", name);
-            des.fail();
-        }
-    }
-    else
-    {
-        des.beginObject();
-        des.read(obj.index);
-        des.read(obj.generation);
-        des.endObject();
-    }
-}
-
-Entity::Entity()
-{
-    this->index = UINT32_MAX;
-    this->generation = UINT32_MAX;
-}
-
-Entity::Entity(uint32_t index, uint32_t generation)
-    : index(index)
-    , generation(generation)
-{
-    // Do nothing.
-}
-
-bool Entity::operator==(const Entity& entity) const
-{
-    return this->index == entity.index && this->generation == entity.generation;
-}
-
-bool Entity::operator!=(const Entity& entity) const
-{
-    return !(*this == entity);
-}
-
-bool Entity::isNull() const
-{
-    return this->index == UINT32_MAX && this->generation == UINT32_MAX;
-}
 
 EntityManager::Iterator::Iterator(const EntityManager& e, const Entity::Mask m)
     : mManager(e)
